SWITCHYARD-970: Break out regex methods into it's own ContextMapper sub-interface
