+++ camel/camel-core/src/main/java/org/switchyard/component/camel/composer/CamelContextMapper.java
@@ -31,14 +31,14 @@ import org.apache.camel.Message;
 import org.switchyard.Context;
 import org.switchyard.Property;
 import org.switchyard.Scope;
-import org.switchyard.component.common.composer.BaseContextMapper;
+import org.switchyard.component.common.composer.BaseRegexContextMapper;
 
 /**
  * CamelContextMapper.
  *
  * @author David Ward &lt;<a href="mailto:dward@jboss.org">dward@jboss.org</a>&gt; (C) 2011 Red Hat Inc.
  */
-public class CamelContextMapper extends BaseContextMapper<Message> {
+public class CamelContextMapper extends BaseRegexContextMapper<Message> {
 
     private static final Scope[] IN_OUT = new Scope[]{IN, OUT};
